add debug place pages
